maybe this works?? re-implemented API methods to use sets for follow(er)s and lists for tls

+++ HW1/src/Twitter/HW2/TwitterHW2Imp.java
@@ -4,83 +4,88 @@ import Twitter.Tweet;
 import Twitter.TwitterAPI;
 import redis.clients.jedis.Jedis;
 
+import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
  * Implements the TwitterAPI Interface using a connection to Redis, implementing Strategy 2 from the HW Doc
- *
+ * <p>
  * The Redis schema contains tweet objects represented as tweet:{tweed_Id} as well as a latestTweetID
  * which is the id of the latest tweet in the system. It also contains a set of users in the system, including
- * users who have tweeted and users who follow or are followed by users who have tweeted. We store a 
+ * users who have tweeted and users who follow or are followed by users who have tweeted. We store a
  * latestUserID which is the id of the latest user in the system. Users are represented as user:{user_Id} with
  * user_Id being the id of the user. Their followers are represented as followers:{user_Id} (a list of people
  * who follow the user) and their timeline is represented as timeline:{user_Id} (a list of tweets posted by this
  * user's follows).
  */
 public class TwitterHW2Imp implements TwitterAPI {
-    private Jedis jedis;
-    private int latestTweetID;
+    private final Jedis jedis;
+
+    public static String LATEST_TWEET_ID_KEY = "latestTweetId";
+    public static String FOLLOWERS_PREFIX = "followers:";
+    public static String FOLLOWS_PREFIX = "follows:";
+    public static String TIMELINE_PREFIX = "timeline:";
+    public static String TWEET_HASH_KEY = "tweet";
 
     public TwitterHW2Imp() {
         this.jedis = new Jedis();
-        this.latestTweetID = 1;
-        jedis.flushAll();
+        jedis.set("latestTweetId", "1");
     }
+
+    /**
+     * Constructs an instance of the TwitterHW2Imp, with the given option to flush what is currently in the db.
+     *
+     * @param flush - a boolean representing if we should flush the db or not. Flush if true, else do nothing.
+     */
+    public TwitterHW2Imp(boolean flush) {
+        this();
+        if (flush) {
+            jedis.flushAll();
+        }
+    }
+
 
     @Override
     public void postTweet(Tweet t) {
-        this.jedis.hset("tweet:" + (latestTweetID + 1), "user_id", t.getUserId()); // add tweet to database with the latest tweet id
-
+        String latestTweetId = getLatestTweetId();
+        // add tweet to database with the latest tweet id in the form userId tweetTxt tweetTS
+        this.jedis.hset(TWEET_HASH_KEY, latestTweetId, t.toString());
         Set<Integer> followers = this.getFollowers(t.getUserId()); // get followers of the user who posted this tweet
-        
         // for each follower, add tweet id to their timeline
         for (Integer follower : followers) {
-            this.jedis.lpush("timeline:" + follower, latestTweetID);
+            this.jedis.lpush(TIMELINE_PREFIX + follower, latestTweetId);
         }
-
-        latestTweetID++; // increment latest tweet id
+        incrementLatestTweetId();
     }
 
     @Override
-    private List<Integer> getTimeline(int userId) {
-        // get the last 10 tweets from the timeline of the user with the given user id
-        List<String> fullTimelineString = this.jedis.hget("user:" + userId, "timeline");
-        // get the last 10 tweets from the timeline
-        List<String> timelineString = fullTimelineString.subList(Math.max(0, fullTimelineString.size() - 10), fullTimelineString.size());
+    public List<Tweet> getTimeline(int userId) {
+        // get latest 10 tweetIDs from jedis, and convert them to ints
+        List<Integer> tweetIds = jedis.lrange(TIMELINE_PREFIX + userId, 0, 10).stream()
+                .map(Integer::parseInt).collect(Collectors.toList());
 
-        // convert the timeline to a list of tweet ids
-        List<Integer> timeline = new ArrayList<>();
-        for (String tweetId : timelineString) {
-            try {
-                timeline.add(Integer.parseInt(tweetId));
-            } catch (NumberFormatException e) {
-                System.out.println("Error parsing tweet id: " + tweetId);
-            }
-        }
-
-        return timeline;
+        // for each tweet, get the tweet hash info, parse it, and convert it to a Tweet object
+        return tweetIds.stream().map(id -> {
+            String tweetInfo = jedis.hget(TWEET_HASH_KEY, Integer.toString(userId));
+            List<String> seperatedTweetInfo = Arrays.asList(tweetInfo.split(" "));
+            return new Tweet(id, userId, seperatedTweetInfo.get(0), Timestamp.valueOf(seperatedTweetInfo.get(1)));
+        }).collect(Collectors.toList());
     }
 
     @Override
     public Set<Integer> getFollowers(int userId) {
-        List<String> followers = this.jedis.hget("user:" + userId, "followers");
-        List<Integer> followerIds = new ArrayList<>();
-
-        for (String follower : followers) {
-            try {
-                followerIds.add(Integer.parseInt(follower));
-            } catch (NumberFormatException e) {
-                System.out.println("Error parsing follower: " + follower);
-            }
-        }
-
-        return followerIds;
+        Set<String> followers = jedis.smembers(FOLLOWERS_PREFIX + userId);
+        return followers.stream().map(Integer::parseInt).collect(Collectors.toSet());
     }
 
     @Override
     public Set<Integer> getFollowees(int userId) {
-        throw new UnsupportedOperationException("We are not doing this for this hw lol");
+        Set<String> follows = jedis.smembers(FOLLOWS_PREFIX + userId);
+        return follows.stream().map(Integer::parseInt).collect(Collectors.toSet());
     }
 
     @Override
@@ -106,29 +111,24 @@ public class TwitterHW2Imp implements TwitterAPI {
 
     @Override
     public void addFollow(int userId, int followId) {
-        // if the user does not already exist, create a followers, follows list, and timeline for them.
-        // if the follow does not already exist, create a followers, follows list, and timeline for them.
-        // add the user to the list of users
         // add the userId to the followId's followers
+        jedis.sadd(FOLLOWERS_PREFIX + followId, Integer.toString(userId));
+
         // add the followID to the userId's follows (**NECESSARY FOR GET TIMELINE**)
-        if (!this.jedis.exists("user:" + userId)) {
-            this.jedis.hset("user:" + userId, "followers", "followers:" + userId);
-            this.jedis.hset("user:" + userId, "follows", "follows:" + userId);
-            this.jedis.hset("user:" + userId, "timeline", "timeline:" + userId);
-            this.jedis.lpush("users", userId);
-        }
-        if (!this.jedis.exists("user:" + followId)) {
-            this.jedis.hset("user:" + followId, "follows", "follows:" + followId);
-            this.jedis.hset("user:" + followId, "followers", "followers:" + followId);
-            this.jedis.hset("user:" + followId, "timeline", "timeline:" + followId);
-            this.jedis.lpush("users", followId);
-        }
-        this.jedis.sadd("followers:" + followId, userId);
-        this.jedis.sadd("follows:" + userId, followId);
+        jedis.sadd(FOLLOWS_PREFIX + userId, Integer.toString(followId));
     }
 
     @Override
     public void closeConnection() {
-        this.jedis.close();
+        jedis.close();
+        jedis.flushAll();
+    }
+
+    private void incrementLatestTweetId() {
+        jedis.incr(LATEST_TWEET_ID_KEY);
+    }
+
+    private String getLatestTweetId() {
+        return jedis.get(LATEST_TWEET_ID_KEY);
     }
 }
